Dead reckoning
Updated CameraController so that is moves based on velocity. Updated ClientManger so that it only sends packets when the player has moved, rotated or enough time has passed. When receiving player data, it only sets the position if the change in position is above cMovementFaultTolerance. During the update loop, local clients are moved based on their velocity