Player names finished
Players can now enter a name. For this, a basic scene system has been added. The first scene consists of one entity with the new NameGetter component. This component gets the name and displays it while it's being entered. When the player presses enter, the player name is set and the old scene is swapped to